+++ optree/registry.py
@@ -23,7 +23,6 @@ import dataclasses
 import functools
 import inspect
 import sys
-import warnings
 from collections import OrderedDict, defaultdict, deque, namedtuple
 from operator import itemgetter, methodcaller
 from threading import Lock
@@ -37,15 +36,11 @@ from typing import (
     Generic,
     Iterable,
     NamedTuple,
-    Sequence,
     Type,
     TypeVar,
     overload,
 )
-from typing_extensions import (
-    TypeAlias,  # Python 3.10+
-    deprecated,  # Python 3.13+
-)
+from typing_extensions import TypeAlias  # Python 3.10+
 
 import optree._C as _C
 from optree.accessor import (
@@ -432,7 +427,7 @@ def register_pytree_node(
     return cls
 
 
-del pytree_node_registry_get
+del pytree_node_registry_get, _add_get
 
 
 CustomTreeNodeType: TypeAlias = Type[CustomTreeNode[T]]
@@ -885,148 +880,3 @@ _DEFAULTDICT_INSERTION_ORDERED_REGISTRY_ENTRY = PyTreeNodeRegistryEntry(
     path_entry_type=MappingEntry,
     kind=PyTreeKind.DEFAULTDICT,
 )
-
-
-####################################################################################################
-
-with warnings.catch_warnings():
-    warnings.filterwarnings('ignore', category=FutureWarning, module=__name__, append=False)
-
-    @deprecated(
-        'The function `_sorted_keys` is deprecated and will be removed in a future version.',
-        category=FutureWarning,
-    )
-    def _sorted_keys(dct: dict[KT, VT], /) -> list[KT]:
-        return total_order_sorted(dct)
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    class KeyPathEntry(NamedTuple):  # pylint: disable=missing-class-docstring
-        key: Any
-
-        def __add__(self, other: object, /) -> KeyPath:
-            if isinstance(other, KeyPathEntry):
-                return KeyPath((self, other))
-            if isinstance(other, KeyPath):
-                return KeyPath((self, *other.keys))
-            return NotImplemented
-
-        def __eq__(self, other: object, /) -> bool:
-            return isinstance(other, self.__class__) and self.key == other.key
-
-        def pprint(self, /) -> str:
-            """Pretty name of the key path entry."""
-            raise NotImplementedError
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    class KeyPath(NamedTuple):  # pylint: disable=missing-class-docstring
-        keys: tuple[KeyPathEntry, ...] = ()
-
-        def __add__(self, other: object, /) -> KeyPath:
-            if isinstance(other, KeyPathEntry):
-                return KeyPath((*self.keys, other))
-            if isinstance(other, KeyPath):
-                return KeyPath(self.keys + other.keys)
-            return NotImplemented
-
-        def __eq__(self, other: object, /) -> bool:
-            return isinstance(other, KeyPath) and self.keys == other.keys
-
-        def pprint(self, /) -> str:
-            """Pretty name of the key path."""
-            if not self.keys:
-                return ' tree root'
-            return ''.join(k.pprint() for k in self.keys)
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    class GetitemKeyPathEntry(KeyPathEntry):
-        """The key path entry class for sequences and dictionaries."""
-
-        def pprint(self, /) -> str:
-            """Pretty name of the key path entry."""
-            return f'[{self.key!r}]'
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    class AttributeKeyPathEntry(KeyPathEntry):
-        """The key path entry class for namedtuples."""
-
-        def pprint(self, /) -> str:
-            """Pretty name of the key path entry."""
-            return f'.{self.key}'
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    class FlattenedKeyPathEntry(KeyPathEntry):  # fallback
-        """The fallback key path entry class."""
-
-        def pprint(self, /) -> str:
-            """Pretty name of the key path entry."""
-            return f'[<flat index {self.key}>]'
-
-    KeyPathHandler = Callable[[Collection[T]], Sequence[KeyPathEntry]]
-    _KEYPATH_REGISTRY: dict[type[Collection], KeyPathHandler] = {}
-
-    @deprecated(
-        'The key path API is deprecated and will be removed in a future version. '
-        'Please use the accessor API instead.',
-        category=FutureWarning,
-    )
-    @_add_get(_KEYPATH_REGISTRY.get)
-    def register_keypaths(
-        cls: type[Collection[T]],
-        /,
-        handler: KeyPathHandler[T],
-    ) -> KeyPathHandler[T]:
-        """Register a key path handler for a custom pytree node type."""
-        if not inspect.isclass(cls):
-            raise TypeError(f'Expected a class, got {cls!r}.')
-        if cls in _KEYPATH_REGISTRY:
-            raise ValueError(f'Key path handler for {cls!r} has already been registered.')
-
-        _KEYPATH_REGISTRY[cls] = handler
-        return handler
-
-    register_keypaths(
-        tuple,
-        lambda tup: list(map(GetitemKeyPathEntry, range(len(tup)))),
-    )
-    register_keypaths(
-        list,
-        lambda lst: list(map(GetitemKeyPathEntry, range(len(lst)))),
-    )
-    register_keypaths(
-        dict,
-        lambda dct: list(map(GetitemKeyPathEntry, _sorted_keys(dct))),  # type: ignore[arg-type]
-    )
-    register_keypaths(
-        OrderedDict,
-        lambda odct: list(map(GetitemKeyPathEntry, odct)),
-    )
-    register_keypaths(
-        defaultdict,
-        lambda ddct: list(map(GetitemKeyPathEntry, _sorted_keys(ddct))),  # type: ignore[arg-type]
-    )
-    register_keypaths(
-        deque,
-        lambda dq: list(map(GetitemKeyPathEntry, range(len(dq)))),
-    )
-
-del _add_get