+++ tests/test_registry.py
@@ -15,7 +15,6 @@
 
 # pylint: disable=missing-function-docstring,invalid-name
 
-import platform
 import re
 import sys
 import weakref
@@ -573,7 +572,7 @@ def test_pytree_node_registry_get_with_invalid_arguments():
     assert optree.register_pytree_node.get(None) == registry
     assert optree.register_pytree_node.get(namespace=GLOBAL_NAMESPACE) == registry
     assert optree.register_pytree_node.get(namedtuple) is registry[namedtuple]  # noqa: PYI024
-    if not (sys.version_info[:2] == (3, 9) and platform.system() == 'Windows'):
+    if sys.version_info[:2] != (3, 9):
         with pytest.raises(TypeError, match='Expected a class or None'):
             optree.register_pytree_node.get(dataclass)
         with pytest.raises(TypeError, match='The namespace must be a string'):