+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
--
+- Remove deprecated key path APIs by [@XuehaiPan](https://github.com/XuehaiPan) in [#195](https://github.com/metaopt/optree/pull/195).
 
 ------